requirements.txt
Flask==2.2.3
tensorflow==2.12.0
numpy==1.23.0
Flask==2.2.3
Werkzeug==2.3.3
tensorflow==2.12.0
Pillow==9.0.0
